BACKEND COMMIT
In this commit, the database is scaffoled by ORM in fist-code, using the
ApplicationDbContext class, and connecting to local dev mssql database.
Added a repository pattern to retrieve data from database, using auto
mapper with viewmodel to retrieve only data needed about the shop
entity.

Now I'm working on the user management, signing and regestering and
using the jwt token to protect the api.